Support KIND_PAD in keylib.py (issue #30)
New batch records will start with an array of constant-width key
suffixes, padded at the end with NUL bytes. KIND_PAD is 0x00, which is
NULL, which tells the key decoder to stop reading more bytes.

This allows us to avoid storing the length of the suffix, or some other
fancier trick.